Bonjour Discovery
Accessories advertise their presence on the network via Bonjour. HAP clients
browse for them and if any are found and determined to be compatible, they are
displayed to the user. Additional data needed for discovery-time metadata is
advertised via a TXT record associated with the Bonjour service. TXT record
keys are intentionally short to minimize network traffic.