Refactor to use unpack1
When you unpack something, the resulting object will always be an array. Often,
this array consists of only single element, which you are interested in (for
example, this is the case for all String Encoding Directives). This is why the
unpack1 has been introduced - It will return the first element directly: