Socket session data
HAP doesn't support cookies but I need to persist state between requests. This
state was being stored as a global but different connections are at different
stages of their lifecycle and shouldn't be sharing state.

Since HAP requests are very long running TCP connections I decided a
way to do this was to get the object_id of the TCP socket and use that as the
key to the session storage.

Not sure how secure this is but will at least let me finish prototyping.

Need to add some clean up so after the connection is closed we clean out the cache.

This will stop
 - A memory leak
 - Recycled object_id from a prior socket connection having access to expired cached data