Switch characteristic to value object
Sometimes when sending a request via the home app the value it sends is
different from the one Siri sends.

For example:
- The home app sends 0 or 1 to turn a fan off or on
- Siri sends false or true to turn a fan off or on.

I want to encapsulate this business logic in value objects so I've decided
characteristics should delegate reading and writing of their values to value
objects for that correspond to their type.