Require Ruby 2.4.1
I have discovered that in Ruby `2.4.0` there appears to be a bit of an issue
with the method `unpack1`.

For example:

```ruby
'8d'.unpack('H*') #=> ["3864"]
'8d'.unpack1('H*') #=> "\x00\x00\x00\x00"
```

However, in Ruby 2.4.1 and future versions, it correctly returns:

```ruby
'8d'.unpack('H*') #=> ["3864"]
'8d'.unpack1('H*') #=> "3864"
```

I *think* this is/was a bug in Ruby `2.4.0` that was fixed in `2.4.1`. This
means RubyHome is currently incompatible with Ruby `2.4.0`.