Expose maximum/minimum/step values of a numeric characteristics (#126)
This exposes the optional Properties of Characteristic Objects in JSON 

- Minimum Value
- Maximum Value
- Step Value

These characteristics are only relevant for characteristics of format
int or float. They allow finer specification of characteristics beyond
the defaults. For example: if someone wants to limit the cooling
temperature of an AC unit `heater_cooler` with no heater between 17C and
31C in 1C steps.